fix: KPI-263 fix json serialize decimals

+++ src/service/calculator/calculator_service.py
@@ -257,7 +257,7 @@ class CalculatorService:
     def calculate_debt_ebitda(self, long_term_debt: float, ebitda: float):
         try:
             debt_ebitda = long_term_debt / ebitda
-            return round(debt_ebitda, 2)
+            return float(round(debt_ebitda, 2))
         except Exception as error:
             self.logger.info(error)
             return "NA"
